Develop

+++ backend/IncaCore/src/main/java/com/Incamar/IncaCore/dtos/vesselItem/VesselItemRes.java
@@ -1,11 +1,13 @@
 package com.Incamar.IncaCore.dtos.vesselItem;
 
+import java.math.BigDecimal;
+
 public record VesselItemRes(
         Long id,
         String name,
         String description,
         String controlType,
-        Integer accumulatedHours,
+        BigDecimal accumulatedHours,
         Integer usefulLifeHours,
         Integer alertHours,
         String materialType,
